+++ osxphotos/bookmark.py
@@ -1,5 +1,7 @@
 """Work with macOS CFURL Bookmarks"""
 
+from __future__ import annotations
+
 import os
 import pathlib